Moved to UDP for the comms

+++ OtterGameState/Private/Otter/GameState/GameState.c
@@ -1,6 +1,6 @@
-#include "GameState.h"
+#include "Otter/GameState/GameState.h"
 
-#include "Player/Player.h"
+#include "Otter/GameState/Player/Player.h"
 
 void game_state_update(PlayerInput* playerInputs, float deltaTime)
 {
@@ -8,8 +8,8 @@ void game_state_update(PlayerInput* playerInputs, float deltaTime)
   {
     if (g_listOfPlayers[i].active && playerInputs[i].actions > 0)
     {
-      float deltaX = playerInputs[i].left * -1 + playerInputs[i].right;
-      float deltaY = playerInputs[i].up * -1 + playerInputs[i].down;
+      float deltaX = (float)(playerInputs[i].left * -1 + playerInputs[i].right);
+      float deltaY = (float)(playerInputs[i].up * -1 + playerInputs[i].down);
 
       g_listOfPlayers[i].x += deltaX * deltaTime * PLAYER_SPEED;
       g_listOfPlayers[i].y += deltaY * deltaTime * PLAYER_SPEED;
